+++ NEWS.md
@@ -1,5 +1,7 @@
 # crochet 2.1.0.9000
 
+* Move test suite to `inst/test-suite` so that it can be used without
+  `--install-tests`.
 
 
 # crochet 2.1.0